Add more types and packets

+++ mclib/protocol/src/packets/base.rs
@@ -1,7 +1,7 @@
 use std::io::Read;
 
 pub trait MCPacket {
-    fn packet_id(&self) -> u8;
+    fn packet_id(&self) -> i32;
     fn pack(&self) -> Vec<u8>;
     fn unpack(src: &mut dyn Read) -> Self;
 }
